requirements.txt
codecov==2.1.12
defusedxml==0.7.1
flake8==3.9.2
httpretty==1.1.4
pluggy==1.0.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyramid_debugtoolbar==4.9
pytest==6.2.5
pytest-cov==2.12.1
PyYAML==5.4.1
twine==3.4.2
-e .
